% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.r
\name{ST}
\alias{ST}
\title{The score testing method for change-plane Cox models similar in Kang et al.(2017).}
\usage{
    ST(data, Gamma, B = 1000, K = 1000, qlb = 0.1, tol = 1e-8)
}
\arguments{
    \item{data}{A list, including \eqn{y} (time response), \eqn{x} (predictors), \eqn{z} (predictors), \eqn{u} (grouping variables), status (censoring indicator).}
    \item{Gamma}{A matrix for initial gamma values. If not given then it will be automatically generated based on the data.}
    \item{B}{A constant. Number of bootstrap samples. Default is 1000.}
    \item{K}{A constant. The number of the initial gamma values. Default is 1000.}
    \item{qlb}{A constant. The lower quantile specified for Z\%*\%gamma.initials. Default is 0.1.}
    \item{tol}{A constant. The precision of the Newton method. Default is \code{1e-8}.}
}
\value{
A list.
\itemize{
    \item TestR - The value of test statistic.
    \item TestB - B values of test statistic obtained from the bootstrap.
    \item Pvalue - The p-value of the test.
    \item time - Running time.
}
}
\references{Kang, S., Lu, W., and Song, R. (2017). Subgroup detection and sample size calculation with proportional hazards regression for survival data. Statistics in medicine, 36(29), 4646-4659.}
\examples{
    n = 100
    p1 = 2
    p2 = 1
    p3 = 3
    alpha = rep(1, p1)
    beta  = rep(1, p2)/2
    gamma = c(1, seq(-1,1,length.out = p3-1)) 
    rho = 0.3
    cenRate = 0.2
    set.seed(100)
    data = generate_cox_data(n, alpha, beta, gamma, rho, cenRate = cenRate)
    fit <- ST(data)
    fit$Pvalue
}