% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WAST.r
\name{WAST}
\alias{WAST}
\title{The weighted-sum-based subgrouping testing method for change-plane Cox models.}
\usage{
    WAST(data, B = 1000, CoreNum = 4, par = TRUE)
}
\arguments{
    \item{data}{A list, including \eqn{y} (time response), \eqn{x} (predictors), \eqn{z} (predictors), \eqn{u} (change-plane variables), status (censoring indicator).}
    \item{B}{A constant. Number of bootstrap samples. Default is 1000.}
    \item{CoreNum}{A constant. Number of processes to be used in parallel operation. Default is 4.}
    \item{par}{A logical value. Whether to conduct the parallel operation. Default is TRUE.}
}
\value{
A list.
\itemize{
    \item alpha_hat - The estimator of alpha under the null hypothesis.
    \item test_statistic - The value of test statistic.
    \item test_statistic_boot - B values of test statistic obtained from the bootstrap.
    \item pval - The p-value of the test.
    \item time - Running time.
}
}
\examples{
    n = 100
    p1 = 2
    p2 = 1
    p3 = 3
    alpha = rep(1, p1)
    beta  = rep(1, p2)/2
    gamma = c(1, seq(-1,1,length.out = p3-1)) 
    rho = 0.3
    cenRate = 0.2
    set.seed(100)
    data = generate_cox_data(n, alpha, beta, gamma, rho, cenRate = cenRate)
    fit <- WAST(data)
    fit$pval
}